move shell_environment to common
