+++ homcc/client/compilation.py
@@ -43,7 +43,7 @@ from homcc.common.messages import (
     Message,
 )
 from homcc.common.statefile import StateFile
-from homcc.server.shell_environment import HostShellEnvironment
+from homcc.common.shell_environment import HostShellEnvironment
 
 logger = logging.getLogger(__name__)
