+++ homcc/common/shell_environment.py
