+++ homcc/server/environment.py
@@ -17,7 +17,7 @@ from homcc.common.messages import CompilationResultMessage, File
 from homcc.server.cache import Cache
 from homcc.server.docker import DockerShellEnvironment
 from homcc.server.schroot import SchrootShellEnvironment
-from homcc.server.shell_environment import HostShellEnvironment, ShellEnvironment
+from homcc.common.shell_environment import HostShellEnvironment, ShellEnvironment
 
 logger = logging.getLogger(__name__)
