+++ tests/server/environment_test.py
@@ -14,7 +14,7 @@ from homcc.common.arguments import Arguments
 from homcc.common.compression import NoCompression
 from homcc.server.cache import Cache
 from homcc.server.environment import ArgumentsExecutionResult, Environment
-from homcc.server.shell_environment import HostShellEnvironment
+from homcc.common.shell_environment import HostShellEnvironment
 
 
 def create_mock_environment(instance_folder: str, mapped_cwd: str) -> Environment:
